backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.1.19
langchain-core==0.2.43
langchain-openai==0.1.25
python-dotenv==1.0.1
pydantic>=2.7.4
